# T&D: Do not deduplicate raw tables