--- a/airbyte-integrations/bases/base-typing-deduping-test/src/main/java/io/airbyte/integrations/base/destination/typing_deduping/BaseSqlGeneratorIntegrationTest.java
+++ b/airbyte-integrations/bases/base-typing-deduping-test/src/main/java/io/airbyte/integrations/base/destination/typing_deduping/BaseSqlGeneratorIntegrationTest.java
@@ -539,7 +539,7 @@ public abstract class BaseSqlGeneratorIntegrationTest<DialectTableDefinition> {
     final List<JsonNode> actualRawRecords = dumpRawTableRecords(streamId);
     final List<JsonNode> actualFinalRecords = dumpFinalTableRecords(streamId, "");
     verifyRecordCounts(
-        1,
+        2,
         actualRawRecords,
         1,
         actualFinalRecords);
@@ -676,10 +676,9 @@ public abstract class BaseSqlGeneratorIntegrationTest<DialectTableDefinition> {
     destinationHandler.execute(sql);
 
     verifyRecordCounts(
-        // We keep the newest raw record per PK
-        7,
+        11,
         dumpRawTableRecords(streamId),
-        5,
+        4,
         dumpFinalTableRecords(streamId, ""));
   }
 
@@ -708,7 +707,7 @@ public abstract class BaseSqlGeneratorIntegrationTest<DialectTableDefinition> {
     destinationHandler.execute(sql);
 
     verifyRecordCounts(
-        1,
+        2,
         dumpRawTableRecords(streamId),
         0,
         dumpFinalTableRecords(streamId, ""));
@@ -745,9 +744,9 @@ public abstract class BaseSqlGeneratorIntegrationTest<DialectTableDefinition> {
     destinationHandler.execute(sql);
 
     verifyRecordCounts(
-        1,
+        2,
         dumpRawTableRecords(streamId),
-        1,
+        0,
         dumpFinalTableRecords(streamId, ""));
   }