--- a/airbyte-integrations/connectors/destination-bigquery/src/test-integration/resources/dat/sync2_cursorchange_expectedrecords_incremental_dedup_raw.jsonl
+++ b/airbyte-integrations/connectors/destination-bigquery/src/test-integration/resources/dat/sync2_cursorchange_expectedrecords_incremental_dedup_raw.jsonl
@@ -1,3 +1,9 @@
+// previous rows in the raw table
+{"_airbyte_extracted_at":"1970-01-01T00:00:01Z", "_airbyte_data": {"id1":1,"id2":200,"old_cursor":0,"_ab_cdc_deleted_at":null,"name":"Alice","address":{"city":"San Francisco","state":"CA"}}}
+{"_airbyte_extracted_at":"1970-01-01T00:00:01Z", "_airbyte_data": {"id1":1,"id2":200,"old_cursor":1,"_ab_cdc_deleted_at":null,"name":"Alice","address":{"city":"Los Angeles","state":"CA"}}}
+{"_airbyte_extracted_at":"1970-01-01T00:00:01Z", "_airbyte_data": {"id1":1,"id2":201,"old_cursor":2,"name":"Bob","address":{"city":"Boston","state":"MA"}}}
+{"_airbyte_extracted_at":"1970-01-01T00:00:02Z", "_airbyte_data": {"id1":1,"id2":201,"updated_at":"2000-01-02T00:00:00Z","_ab_cdc_deleted_at":null,"name":"Bob","address":{"city":"New York","state":"NY"}}}
+// test specific rows
 {"_airbyte_extracted_at": "1970-01-01T00:00:02Z", "_airbyte_data": {"id1": 1, "id2": 200, "updated_at": "2000-01-02T00:00:00Z", "_ab_cdc_deleted_at": null, "name": "Alice", "address": {"city": "Seattle", "state": "WA"}}}
 {"_airbyte_extracted_at": "1970-01-01T00:00:02Z", "_airbyte_data": {"id1": 1, "id2": 201, "updated_at": "2000-01-02T00:01:00Z", "_ab_cdc_deleted_at": "1970-01-01T00:00:00Z"}}
 // Charlie wasn't reemitted in sync2. This record still has an old_cursor value.