--- a/airbyte-integrations/connectors/destination-bigquery/src/test-integration/resources/dat/sync2_expectedrecords_incremental_dedup_raw.jsonl
+++ b/airbyte-integrations/connectors/destination-bigquery/src/test-integration/resources/dat/sync2_expectedrecords_incremental_dedup_raw.jsonl
@@ -1,3 +1,9 @@
+// Previous Syncs
+{"_airbyte_extracted_at":"1970-01-01T00:00:01Z", "_airbyte_data": {"id1":1,"id2":200,"updated_at":"2000-01-01T00:00:00Z","_ab_cdc_deleted_at":null,"name":"Alice","address":{"city":"San Francisco","state":"CA"}}}
+{"_airbyte_extracted_at":"1970-01-01T00:00:01Z", "_airbyte_data": {"id1":1,"id2":200,"updated_at":"2000-01-01T00:01:00Z","_ab_cdc_deleted_at":null,"name":"Alice","address":{"city":"Los Angeles","state":"CA"}}}
+{"_airbyte_extracted_at":"1970-01-01T00:00:01Z", "_airbyte_data": {"id1":1,"id2":201,"updated_at":"2000-01-01T00:02:00Z","name":"Bob","address":{"city":"Boston","state":"MA"}}}
+{"_airbyte_extracted_at":"1970-01-01T00:00:02Z", "_airbyte_data": {"id1":1,"id2":201,"updated_at":"2000-01-02T00:00:00Z","_ab_cdc_deleted_at":null,"name":"Bob","address":{"city":"New York","state":"NY"}}}
+// Orignal
 {"_airbyte_extracted_at": "1970-01-01T00:00:02Z", "_airbyte_data": {"id1": 1, "id2": 200, "updated_at": "2000-01-02T00:00:00Z", "_ab_cdc_deleted_at": null, "name": "Alice", "address": {"city": "Seattle", "state": "WA"}}}
 // Keep the record that deleted Bob, but delete the other records associated with id=(1, 201)
 {"_airbyte_extracted_at": "1970-01-01T00:00:02Z", "_airbyte_data": {"id1": 1, "id2": 201, "updated_at": "2000-01-02T00:01:00Z", "_ab_cdc_deleted_at": "1970-01-01T00:00:00Z"}}