Closes #30710

I've confirmed that V1 normalization does not dedupe the raw tables, meaning that it keeps all entries for historic syncs, spending space but reducing compute expense:

```sql
SELECT COUNT(*) from v1_staging_inserts.products; -- 100
SELECT COUNT(*) from v1_staging_inserts._airbyte_raw_products; -- 10300
SELECT * FROM v1_staging_inserts._airbyte_raw_products WHERE CAST(JSON_EXTRACT_SCALAR(_airbyte_data, '$.id') AS INT) IN (1) ORDER BY _airbyte_emitted_at DESC; -- 103 entries
```

## Trials

Publishing dev images to try in the performance workspace... `--name=destination-bigquery --name=destination-snowflake` - https://github.com/airbytehq/airbyte/actions/runs/6304657674

* [airbyte/destination-bigquery:2.1.0-dev.14f54796e1](https://hub.docker.com/r/airbyte/destination-bigquery/tags)
* [airbyte/destination-snowflake:3.2.0-dev.14f54796e1](https://hub.docker.com/r/airbyte/destination-snowflake/tags)

### BigQuery

Before (Users Table) ([connection](https://cloud.airbyte.com/workspaces/b61bc266-ef3c-460c-af2f-f70da4a2993c/connections/31331a3d-d8e5-4c9a-b243-e8b6b2a681dd/job-history)): 
```

2023-09-25 15:00:22 destination > INFO i.a.i.d.b.t.BigQueryDestinationHandler(execute):79 Root-level job 41e52714-450b-4815-99e8-a220b2787d92 completed in 129618 ms; processed 29127233754 bytes; billed for 29129441280 bytes
2023-09-25 15:00:24 destination > INFO i.a.i.d.b.t.BigQueryDestinationHandler(lambda$execute$1):101 Child sql BEGIN TRANSACTION completed in 90 ms; processed 0 bytes; billed for 0 bytes
2023-09-25 15:00:24 destination > INFO i.a.i.d.b.t.BigQueryDestinationHandler(lambda$execute$1):101 Child sql SELECT STRUCT<int64>(( SELECT COUNT(1) FROM `destinationsv2-perf-testing`.`airbyte_internal`.`v2_sta... completed in 1120 ms; processed 5373675012 bytes; billed for 5373952000 bytes</int64>
2023-09-25 15:00:24 destination > INFO i.a.i.d.b.t.BigQueryDestinationHandler(lambda$execute$1):101 Child sql INSERT INTO `destinationsv2-perf-testing`.`v2_staging_inserts`.`users` ( `address`, `occupation`, `g... completed in 93491 ms; processed 5835658210 bytes; billed for 5836374016 bytes
2023-09-25 15:00:24 destination > INFO i.a.i.d.b.t.BigQueryDestinationHandler(lambda$execute$1):101 Child sql DELETE FROM `destinationsv2-perf-testing`.`v2_staging_inserts`.`users` WHERE `_airbyte_raw_id` IN ( ... completed in 5962 ms; processed 4651362695 bytes; billed for 4651483136 bytes
2023-09-25 15:00:24 destination > INFO i.a.i.d.b.t.BigQueryDestinationHandler(lambda$execute$1):101 Child sql DELETE FROM `destinationsv2-perf-testing`.`airbyte_internal`.`v2_staging_inserts_raw__stream_users` ... completed in 13315 ms; processed 7456285373 bytes; billed for 7456423936 bytes
2023-09-25 15:00:24 destination > INFO i.a.i.d.b.t.BigQueryDestinationHandler(lambda$execute$1):101 Child sql UPDATE `destinationsv2-perf-testing`.`airbyte_internal`.`v2_staging_inserts_raw__stream_users` SET `... completed in 9777 ms; processed 5810252464 bytes; billed for 5811208192 bytes
2023-09-25 15:00:24 destination > INFO i.a.i.d.b.t.BigQueryDestinationHandler(lambda$execute$1):101 Child sql COMMIT TRANSACTION completed in 545 ms; processed 0 bytes; billed for 0 bytes
```

Roughly 25% of the bill is attributed to deduplication of the raw table! 

After (Users Table):
```
2023-09-26 15:56:34 destination > INFO i.a.i.d.b.t.BigQueryDestinationHandler(execute):79 Root-level job f4b66baf-2636-468d-afa7-78a0d7a8bac7 completed in 121924 ms; processed 21622118759 bytes; billed for 21624782848 bytes
2023-09-26 15:56:36 destination > INFO i.a.i.d.b.t.BigQueryDestinationHandler(lambda$execute$1):101 Child sql BEGIN TRANSACTION completed in 70 ms; processed 0 bytes; billed for 0 bytes
2023-09-26 15:56:36 destination > INFO i.a.i.d.b.t.BigQueryDestinationHandler(lambda$execute$1):101 Child sql SELECT STRUCT<int64>(( SELECT COUNT(1) FROM `destinationsv2-perf-testing`.`airbyte_internal`.`v2_sta... completed in 6600 ms; processed 5350252073 bytes; billed for 5350883328 bytes</int64>
2023-09-26 15:56:36 destination > INFO i.a.i.d.b.t.BigQueryDestinationHandler(lambda$execute$1):101 Child sql INSERT INTO `destinationsv2-perf-testing`.`v2_staging_inserts`.`users` ( `address`, `occupation`, `g... completed in 89854 ms; processed 5810252073 bytes; billed for 5811208192 bytes
2023-09-26 15:56:36 destination > INFO i.a.i.d.b.t.BigQueryDestinationHandler(lambda$execute$1):101 Child sql DELETE FROM `destinationsv2-perf-testing`.`v2_staging_inserts`.`users` WHERE `_airbyte_raw_id` IN ( ... completed in 6997 ms; processed 4651362540 bytes; billed for 4651483136 bytes
2023-09-26 15:56:36 destination > INFO i.a.i.d.b.t.BigQueryDestinationHandler(lambda$execute$1):101 Child sql UPDATE `destinationsv2-perf-testing`.`airbyte_internal`.`v2_staging_inserts_raw__stream_users` SET `... completed in 14594 ms; processed 5810252073 bytes; billed for 5811208192 bytes
2023-09-26 15:56:36 destination > INFO i.a.i.d.b.t.BigQueryDestinationHandler(lambda$execute$1):101 Child sql COMMIT TRANSACTION completed in 330 ms; processed 0 bytes; billed for 0 bytes
```

### Snowflake

Extracted from lines like
```
2023-09-25 05:03:25 destination > INFO i.a.i.d.s.t.SnowflakeDestinationHandler(execute):99 Sql fa2c4a72-6ce0-4127-bb70-bce55bda68b1 completed in 118456 ms
```

Before (Users Table): 
* 116226 ms
* 114263 ms
* 118456 ms

After (Users Table):
* 102758 ms
* 104829 ms
* 103056 ms

It's ~10% bit faster... but we don't have good data on the _cost_ at this time.  This also probably not a big enough dataset to know for sure.  But, it's an improvement either way!