@airbyte-oss-build-runner 2023-09-25
## <img src="https://raw.githubusercontent.com/airbytehq/airbyte/14f54796e1d16256e7029b07ba03b6666320b37c/airbyte-integrations/connectors/destination-bigquery/icon.svg" width="40" height="40"> destination-bigquery test report (commit [`14f54796e1`](https://github.com/airbytehq/airbyte/pull/30742/commits/14f54796e1d16256e7029b07ba03b6666320b37c)) - ❌

⏲️  Total pipeline duration: 07mn07s 

| Step                                                                        | Result   |
|:----------------------------------------------------------------------------|:---------|
| Build connector tar                                                         | ✅        |
| Build destination-bigquery docker image for platform linux/x86_64           | ✅        |
| Java Connector Unit Tests                                                   | ✅        |
| Java Connector Integration Tests                                            | ❌        |
| Validate airbyte-integrations/connectors/destination-bigquery/metadata.yaml | ✅        |
| Connector version semver check                                              | ✅        |
| Connector version increment check                                           | ✅        |
| QA checks                                                                   | ❌        |

🔗 [View the logs here](https://storage.cloud.google.com/airbyte-ci-reports-multi/airbyte-ci/connectors/test/pull_request/evan_no-dedpue-raw-table/1695674468/14f54796e1d16256e7029b07ba03b6666320b37c/destination-bigquery/2.1.0/output.html)

☁️ [View runs for commit in Dagger Cloud](https://alpha.dagger.cloud/changeByPipelines?filter=dagger.io/git.ref:14f54796e1d16256e7029b07ba03b6666320b37c)

*Please note that tests are only run on PR ready for review. Please set your PR to draft mode to not flood the CI engine and upstream service on following commits.*
**You can run the same pipeline locally on this branch with the [airbyte-ci](https://github.com/airbytehq/airbyte/blob/master/airbyte-ci/connectors/pipelines/README.md) tool with the following command**
```bash
airbyte-ci connectors --name=destination-bigquery test
```